move cross-val helpers to their own module
